refactor(psm)!: change Percent type from float to int (#52)
BREAKING CHANGE: Percent struct now uses int instead of float for
channel values.
- Modified Percent struct to use int instead of float for channel values
- Updated adjustChannels to handle int-based percentages with explicit
casting
- No functional changes since percentages were already being used as
whole numbers

+++ src/psm/include/psm/percent.hpp
@@ -3,9 +3,9 @@
 namespace psm {
 
 struct Percent {
-  float channel0_;
-  float channel1_;
-  float channel2_;
+  int channel0_;
+  int channel1_;
+  int channel2_;
 };
 
 }  // namespace psm
